test: remove Projects Under Test
