--- src/tests/ProjectUnderTest.NetCore/MyClass2.cs
@@ -1,12 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectUnderTest.NetCore
-{
-    internal class MyClass2
-    {
-    }
-}
