--- src/tests/ProjectUnderTest.NetCore/MyClass6.cs
@@ -1,11 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProjectUnderTest.NetCore;
-
-internal class MyClass6
-{
-}
