--- src/tests/ProjectUnderTest.NetCore/Properties/AssemblyInfo.cs
@@ -1,6 +1,0 @@
-using System.Reflection;
-
-[assembly: AssemblyDescription("A .NET tool that facilitates development workflows.")]
-[assembly: AssemblyCopyright("Copyright © FlashOWare 2023")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
