--- src/tests/ProjectUnderTest.NetFx/Properties/AssemblyInfo.cs
@@ -1,18 +1,0 @@
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("ProjectUnderTest.NetFx")]
-[assembly: AssemblyDescription("A .NET tool that facilitates development workflows.")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("ProjectUnderTest.NetFx")]
-[assembly: AssemblyCopyright("Copyright © FlashOWare 2023")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-[assembly: ComVisible(false)]
-
-[assembly: Guid("efa38915-eba8-41b8-9050-205e2b350a4d")]
-
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
